# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/exam/start": {
      "post": {
        "description": "Select 1 skill and optionally difficulty levels. System will generate 20 randomized questions for that skill.",
        "operationId": "ExamController_startExam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartExamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Start an exam",
        "tags": [
          "Exam"
        ]
      }
    },
    "/exam/submit": {
      "post": {
        "description": "Submit answers for all questions. System will calculate score and assign skill level.",
        "operationId": "ExamController_submitExam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitExamDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Submit exam answers",
        "tags": [
          "Exam"
        ]
      }
    },
    "/exam/my-tests": {
      "get": {
        "description": "Get all tests taken by the authenticated user.",
        "operationId": "ExamController_getMyTests",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get user's exam history",
        "tags": [
          "Exam"
        ]
      }
    },
    "/exam/my-tests/{testId}": {
      "get": {
        "description": "Get detailed results of a specific test including all answers and skill-level assessments.",
        "operationId": "ExamController_getTestDetails",
        "parameters": [
          {
            "name": "testId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get test details",
        "tags": [
          "Exam"
        ]
      }
    }
  },
  "info": {
    "title": "AI Recruit Documentation",
    "description": "This is the API Docs for using internally.",
    "version": "1.0",
    "contact": {},
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local server 1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "StartExamDto": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174001",
            "description": "Select 1 skill to test"
          },
          "difficultyLevels": {
            "example": [
              "easy",
              "medium"
            ],
            "description": "Optional: Filter by difficulty levels. If not provided, random mix of all difficulties",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "skillId"
        ]
      },
      "AnswerDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "chosenAnswer": {
            "type": "string",
            "example": "A programming language"
          }
        },
        "required": [
          "questionId",
          "chosenAnswer"
        ]
      },
      "SubmitExamDto": {
        "type": "object",
        "properties": {
          "userTestId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            }
          }
        },
        "required": [
          "userTestId",
          "answers"
        ]
      }
    }
  },
...
}